Added home page w/ calendar, daily task list, button to view all. Updated registration. Next to do

+++ Daily-Wins/ViewModels/ToDoListViewViewModel.swift
@@ -5,6 +5,7 @@
 //  Created by Eric Kim on 6/24/24.
 //
 
+import FirebaseAuth
 import FirebaseFirestore
 import Foundation
 
@@ -20,13 +21,26 @@ class ToDoListViewViewModel: ObservableObject {
     }
     
     // delete to do list item
-    func delete(id: String) {
+    /*func delete(id: String) {
         let db = Firestore.firestore()
         
         db.collection("users")
             .document(userId)
             .collection("todos")
             .document(id)
+            .delete()
+    }*/
+    
+    func deleteItem(item: ToDoListItem) {
+        guard let uid = Auth.auth().currentUser?.uid else {
+            return
+        }
+        
+        let db = Firestore.firestore()
+        db.collection("users")
+            .document(uid)
+            .collection("todos")
+            .document(item.id)
             .delete()
     }
 }
